Fix confusion between ptrdiff_t and int.
2004-11-01  Morten Welinder  <terra@gnome.org>

	* ms-escher.c (ms_escher_get_data): Fix confusion between
	ptrdiff_t and int.

+++ plugins/excel/ms-escher.c
@@ -223,7 +223,9 @@ ms_escher_get_data (MSEscherState *state,
 
 		/* Copy back stub */
 		memcpy (tmp, res, num_bytes - (tmp-buffer));
-		d (1, printf ("record %d) add %d bytes;\n", ++counter, num_bytes - (tmp-buffer)););
+		d (1, printf ("record %d) add %d bytes;\n",
+			      ++counter,
+			      num_bytes - (int)(tmp-buffer)););
 		return buffer;
 	}
